Removed conftest.py
